% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-dpc-block.R
\name{admiral_dpc_block}
\alias{admiral_dpc_block}
\title{Admiral Block}
\usage{
admiral_dpc_block(data, ...)
}
\arguments{
\item{data}{Block input data}

\item{...}{Further (metadata) attributes}
}
\description{
Admiral Block
}
