% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code_field}
\alias{code_field}
\alias{new_code_field}
\alias{validate_field.code_field}
\title{Code field}
\usage{
new_code_field(value = character(), ...)

\method{validate_field}{code_field}(x)
}
\arguments{
\item{value}{Initial code.}

\item{...}{Ignored.}

\item{x}{Object.}
}
\description{
Code field
}
