% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-preview.R
\name{col_type_label}
\alias{col_type_label}
\title{Column Type Label}
\usage{
col_type_label(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A character string like "\if{html}{\out{<chr>}}", "\if{html}{\out{<int>}}", "\if{html}{\out{<dbl>}}", etc.
}
\description{
Returns tibble-style type labels for column types.
}
\keyword{internal}
