% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-bms-blocks.R
\name{new_asfactor_block}
\alias{new_asfactor_block}
\alias{asfactor_block}
\alias{demo_data_block}
\alias{demo_join_block}
\alias{demo_arrange_block}
\alias{demo_group_by_block}
\alias{demo_filter_block_1}
\alias{demo_filter_block_2}
\alias{demo_summarize_block}
\title{As factor block}
\usage{
new_asfactor_block(data, column = "VISIT", ...)

asfactor_block(data, ...)

demo_data_block(...)

demo_join_block(data, ...)

demo_arrange_block(data, ...)

demo_group_by_block(data, ...)

demo_filter_block_1(data, ...)

demo_filter_block_2(data, ...)

demo_summarize_block(data, ...)
}
\arguments{
\item{data}{Tabular data in which to select some columns.}

\item{column}{Column to apply the operation on.}

\item{...}{Any other params. TO DO}
}
\description{
Useful for the BMS demo app
}
