% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-select.R
\name{fb_select_dep}
\alias{fb_select_dep}
\title{HTML dependency for the fb-select input binding}
\usage{
fb_select_dep()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
HTML dependency for the fb-select input binding
}
\keyword{internal}
