% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-expr-block.R
\name{filter_expr_block}
\alias{filter_expr_block}
\title{Filter Expression Block}
\usage{
filter_expr_block(data, ...)
}
\arguments{
\item{data}{Block input data}

\item{...}{Further (metadata) attributes}
}
\description{
Filter Expression Block
}
