% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-block-base.R
\name{function-block-base}
\alias{function-block-base}
\title{Shared utilities for function blocks}
\description{
Internal functions used by function_block, function_var_block, and
function_xy_block to reduce code duplication.
}
\keyword{internal}
