% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-summary-block.R
\name{new_broom_summary_block}
\alias{new_broom_summary_block}
\title{Broom Summary Block}
\usage{
new_broom_summary_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_transform_block]{blockr.core::new_transform_block()}}}
}
\value{
A transform block object of class \code{broom_summary_block}.
}
\description{
Creates an adaptive HTML summary display for ANY broom-compatible model.
Shows coefficients, model statistics, and influence diagnostics in a
responsive column layout that adapts based on which broom methods work.
}
\details{
Works with lm, glm, lmer, rlm, coxph, gam, and 100+ other model types
supported by broom.
}
\examples{
if (interactive()) {
  library(blockr.core)
  model <- lm(mpg ~ cyl + hp + wt, data = mtcars)
  serve(new_broom_summary_block(), list(data = model))
}

}
