% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-bms-blocks.R
\name{new_plot_block}
\alias{new_plot_block}
\alias{plot_block}
\alias{new_ggiraph_block}
\alias{ggiraph_block}
\alias{server_output.ggiraph_block}
\alias{uiOutputBlock.ggiraph_block}
\title{Monolithic ggplot block}
\usage{
new_plot_block(
  data,
  plot_opts = list(colors = c("blue", "red"), point_size = 3, title = "Plot title", theme
    = c("theme_minimal", "theme_gray", "theme_linedraw", "theme_dark", "theme_light",
    "theme_classic", "theme_void", "theme_bw"), x_lab = "X axis label", y_lab =
    "Y axis label", errors = list(show = FALSE, ymin = character(), ymax = character()),
    lines = list(show = FALSE, group = character(), color = character())),
  ...
)

plot_block(data, ...)

new_ggiraph_block(
  data,
  plot_opts = list(colors = c("blue", "red"), point_size = 3, title = "Plot title", theme
    = c("theme_minimal", "theme_gray", "theme_linedraw", "theme_dark", "theme_light",
    "theme_classic", "theme_void", "theme_bw"), x_lab = "X axis label", y_lab =
    "Y axis label", errors = list(show = TRUE, ymin = character(), ymax = character()),
    lines = list(show = TRUE, group = character(), color = character())),
  ...
)

ggiraph_block(data, ...)

\method{server_output}{ggiraph_block}(x, result, output)

\method{uiOutputBlock}{ggiraph_block}(x, ns)
}
\arguments{
\item{data}{Tabular data in which to select some columns.}

\item{plot_opts}{List containing options for ggplot (color, ...).}

\item{...}{Any other params. TO DO}

\item{x}{Object for which to generate UI components}

\item{result}{Block result}

\item{output}{Shiny output}

\item{ns}{Output namespace}
}
\description{
Used with cdisc data. See \link{demo_data_block}.

To use with \link{demo_data_block} as data input.
}
