% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-expr-block.R
\name{summarize_expr_block}
\alias{summarize_expr_block}
\title{Summarize By Block}
\usage{
summarize_expr_block(data, ...)
}
\arguments{
\item{data}{Block input data}

\item{...}{Further (metadata) attributes}
}
\description{
Summarize By Block
}
