% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code_block}
\alias{code_block}
\alias{new_code_transform_block}
\alias{code_transform_block}
\alias{new_code_plot_block}
\alias{code_plot_block}
\title{Code Block}
\usage{
new_code_transform_block(data, ...)

code_transform_block(data, ...)

new_code_plot_block(data, ...)

code_plot_block(data, ...)
}
\arguments{
\item{data}{Dataset.}

\item{...}{Ignored.}
}
\description{
Code Block
}
